+++ autobump-core/src/main/java/com/github/autobump/core/model/usecases/PullRequestUseCase.java
@@ -1,0 +1,37 @@
+package com.github.autobump.core.model.usecases;
+
+import com.github.autobump.core.model.Bump;
+import com.github.autobump.core.model.GitClient;
+import com.github.autobump.core.model.GitProvider;
+import com.github.autobump.core.model.PullRequest;
+import com.github.autobump.core.model.UrlHelper;
+import com.github.autobump.core.model.Workspace;
+
+import java.net.URI;
+
+public class PullRequestUseCase {
+    private final GitProvider gitProvider;
+    private final GitClient gitClient;
+    private final UrlHelper urlHelper;
+
+    public PullRequestUseCase(GitProvider gitProvider, GitClient gitClient, UrlHelper urlHelper) {
+        this.gitProvider = gitProvider;
+        this.gitClient = gitClient;
+        this.urlHelper = urlHelper;
+    }
+
+    private void doPullRequest(PullRequest pullRequest) {
+        gitProvider.makePullRequest(pullRequest);
+    }
+
+    public void makeAndExecutePullRequest(Workspace workspace, Bump bump, URI uri) {
+        var commitResult = gitClient.commitToNewBranch(workspace, bump);
+        PullRequest pullRequest = PullRequest.builder()
+                .branchName(commitResult.getBranchName())
+                .title(commitResult.getCommitMessage())
+                .repoName(urlHelper.getRepoName(uri.toString()))
+                .repoOwner(urlHelper.getOwnerName(uri.toString()))
+                .build();
+        doPullRequest(pullRequest);
+    }
+}
